DS Assignment

+++ Quadretic.cpp
@@ -6,18 +6,18 @@ int main()
     int a, b, c;
     cin >> a >> b >> c;
 
-    int d = (b * b) - (4 * a * c);
+    double d = (b * b) - (4 * a * c);
 
     if (d > 0)
     {
-        int r1 = (-b + (sqrt(d))) / (2 * a);
+        double r1 = (-b + (sqrt(d))) / (2 * a);
         cout << "First root : " << r1 << endl;
-        int r2 = (-b - (sqrt(d))) / (2 * a);
+        double r2 = (-b - (sqrt(d))) / (2 * a);
         cout << "Second root : " << r2 << endl;
     }
     else if (d == 0)
     {
-        int r = -b / (2 * a);
+        double r = -b / (2 * a);
         cout << "Unique root : " << r << endl;
     }
     else
